Changed comparison counter from int to long
git-svn-id: https://pmd.svn.sourceforge.net/svnroot/pmd/trunk@1638 51baf565-9d33-0410-a72c-fc3788e3496d

+++ pmd/src/net/sourceforge/pmd/cpd/CPDListener.java
@@ -4,5 +4,5 @@ import java.io.File;
 
 public interface CPDListener {
     void addedFile(int fileCount, File file);
-    void comparisonCountUpdate(int comparisons);
+    void comparisonCountUpdate(long comparisons);
 }
